requirements.txt
fastapi
uvicorn[standard]
newspaper3k
spacy>=3.7
transformers>=4.40
torch          # CPU is fine; CUDA auto-enabled if available
pycountry
streamlit      # for the optional UI
